Function in KDTree

+++ DS-Project/DS-Project/component.cpp
@@ -1,5 +1,5 @@
 #include "component.h"
-void component::Push(int x, int y, string name)
+void component::Push(float x, float y, string name)
 {
 	this->x = x;
 	this->y = y;
